requirements.txt
fastapi==0.68.1
uvicorn==0.16.0 
python-dotenv==0.19.2 
mailer==0.8.1 
fastapi_mailman==0.0.3 